requirements.txt
django
MySQL-python
lxml
python-dateutil
django-pure-pagination